+++ .swiftlint.yml
@@ -1,2 +1,3 @@
 excluded:
   - Carthage
+  - .build
