+++ Tests/SwiftQueueTests/TestUtils.swift
@@ -77,11 +77,11 @@ class PersisterTracker: UserDefaultsPersister {
     var restoreQueueName = ""
 
     var putQueueName: [String] = [String]()
-    var putTaskId: [String] = [String]()
+    var putJobUUID: [String] = [String]()
     var putData: [String] = [String]()
 
     var removeQueueName: [String] = [String]()
-    var removeJobId: [String] = [String]()
+    var removeJobUUID: [String] = [String]()
 
     override func restore(queueName: String) -> [String] {
         restoreQueueName = queueName
@@ -90,14 +90,14 @@ class PersisterTracker: UserDefaultsPersister {
 
     override func put(queueName: String, taskId: String, data: String) {
         putQueueName.append(queueName)
-        putTaskId.append(taskId)
+        putJobUUID.append(taskId)
         putData.append(data)
         super.put(queueName: queueName, taskId: taskId, data: data)
     }
 
     override func remove(queueName: String, taskId: String) {
         removeQueueName.append(queueName)
-        removeJobId.append(taskId)
+        removeJobUUID.append(taskId)
         super.remove(queueName: queueName, taskId: taskId)
     }
 }
